Rename context to contexts for future contexts that will be created
