+++ front-end/src/app/[lang]/layout.js
@@ -2,7 +2,7 @@ import "../globals.css";
 import { Inter } from "next/font/google";
 import Navigation from "@/components/Navigation";
 import { getDictionary } from "./dictionaries";
-import { LanguageProvider } from "@/context/LanguageContext";
+import { LanguageProvider } from "@/contexts/LanguageContext";
 
 const inter = Inter({ subsets: ["latin"] });
